Learning Algorithm #0
Gone horribly wrong, and unfinished. It seems the weighed randomization is beyond my capabilities as of now. The idea would have been that a green colour square, would try random actions, (it could only go in 4 directions), and the randomized actions it took would be "recorded" which would modify the weights, should it succeed in reaching the flag i placed for it.  It never got close to the flag. Then i tried to add some weights to it manually, but it then i realized it doesn't work in weighing the actions, and ends up completely randomized. Unfortunate, but the idea might just work if i wasn't so bad at coding